watchdog: Remove log-only/debug taboo logic, always kill forbidden apps
- WatchService.cpp/h: Remove all log-only and debug conditional logic for taboo (forbidden) process control
- Taboo logic now always kills forbidden apps if enabled in ini
- Refactor: async QtConcurrent, clean config-driven logic
- CMakeLists: Add QtConcurrent for build
- DEPRECATED: startTerminatedModules() kept for compatibility only

Config:
- watchdog.ini.mac.in: Only Terminal enabled for debug, recommended apps in comment
- watchdog.ini.linux.in/.win.in: Add log_only keys as comment for reference
- user.ini: Enable taboo logic by default

docs: Add taboo-apps-analysis.md, update burn-in-protection.md for clarity

misc: Fix minor warnings in SplashScreen, PaymentService, DialupConnectionWindow

+++ src/modules/SDK/PaymentProcessor/src/Scripting/PaymentService.cpp
@@ -468,7 +468,7 @@ QVariantMap PaymentService::calculateLimits(const QString &aAmount, bool aFixedA
     }
 
     // Округлим результат после 2ого знака чтобы не было расхождений в копейках
-    maxAmountAll = qRound64(maxAmountAll * 100) / 100.0;
+    maxAmountAll = static_cast<double>(qRound64(maxAmountAll * 100)) / 100.0;
 
     // Если провайдер требует округления - применим округление к пограничным значениям
     if (provider.processor.rounding) {
